move helper classes to its own package
