+++ duel/src/main/java/me/hsgamer/flexegames/game/duel/DuelExtension.java
@@ -5,7 +5,7 @@ import me.hsgamer.flexegames.api.extension.SingleGameExtension;
 import me.hsgamer.flexegames.api.game.Game;
 import me.hsgamer.flexegames.game.duel.manager.DuelWorldManager;
 import me.hsgamer.flexegames.util.ConfigGeneratorUtil;
-import me.hsgamer.flexegames.util.kit.GameKitManager;
+import me.hsgamer.flexegames.helper.kit.GameKitManager;
 
 import java.util.ArrayList;
 import java.util.List;
