+++ duel/src/main/java/me/hsgamer/flexegames/game/duel/DuelPropertyEditor.java
@@ -2,9 +2,9 @@ package me.hsgamer.flexegames.game.duel;
 
 import me.hsgamer.flexegames.builder.ItemBuilder;
 import me.hsgamer.flexegames.util.ItemUtil;
-import me.hsgamer.flexegames.util.editor.PropertySingleValueButton;
-import me.hsgamer.flexegames.util.editor.PropertyValueButton;
-import me.hsgamer.flexegames.util.editor.SimplePropertyEditor;
+import me.hsgamer.flexegames.helper.editor.PropertySingleValueButton;
+import me.hsgamer.flexegames.helper.editor.PropertyValueButton;
+import me.hsgamer.flexegames.helper.editor.SimplePropertyEditor;
 import net.kyori.adventure.text.Component;
 import net.minestom.server.item.ItemStack;
