+++ main/src/main/java/me/hsgamer/flexegames/helper/editor/PropertyEditor.java
@@ -1,4 +1,4 @@
-package me.hsgamer.flexegames.util.editor;
+package me.hsgamer.flexegames.helper.editor;
 
 import me.hsgamer.flexegames.api.property.GamePropertyMap;
 import me.hsgamer.hscore.minecraft.gui.advanced.AdvancedButtonMap;
