+++ main/src/main/java/me/hsgamer/flexegames/helper/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Contains the helper classes to easily create the game
+ */
+package me.hsgamer.flexegames.helper;
