+++ main/src/main/java/me/hsgamer/flexegames/lobby/Lobby.java
@@ -11,6 +11,7 @@ import me.hsgamer.flexegames.builder.InstanceModifierBuilder;
 import me.hsgamer.flexegames.builder.ItemBuilder;
 import me.hsgamer.flexegames.feature.arena.DescriptionFeature;
 import me.hsgamer.flexegames.feature.arena.JoinFeature;
+import me.hsgamer.flexegames.helper.hotbar.HotbarItemsHelper;
 import me.hsgamer.flexegames.manager.GameArenaManager;
 import me.hsgamer.flexegames.manager.ReplacementManager;
 import me.hsgamer.flexegames.util.*;
