+++ pve/src/main/java/me/hsgamer/flexegames/game/pve/PveExtension.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import me.hsgamer.flexegames.api.extension.SingleGameExtension;
 import me.hsgamer.flexegames.api.game.Game;
 import me.hsgamer.flexegames.util.ConfigGeneratorUtil;
-import me.hsgamer.flexegames.util.kit.GameKitManager;
+import me.hsgamer.flexegames.helper.kit.GameKitManager;
 
 import java.util.ArrayList;
 import java.util.List;
